Arrojar excepción en PacienteService operación eliminar. Borrar lógica de validación previa en PacienteController (eliminar)

+++ src/main/java/com/dh/clinica/service/Impl/PacienteService.java
@@ -1,6 +1,7 @@
 package com.dh.clinica.service.Impl;
 
 import com.dh.clinica.entity.Paciente;
+import com.dh.clinica.exception.ResourceNotFoundException;
 import com.dh.clinica.repository.IPacienteRepository;
 import com.dh.clinica.service.IPacienteService;
 import org.springframework.stereotype.Service;
@@ -38,7 +39,13 @@ public class PacienteService implements IPacienteService {
 
     @Override
     public void eliminarPaciente(Integer id) {
-        pacienteRepository.deleteById(id);
+        Optional<Paciente> pacienteEncontrado = pacienteRepository.findById(id);
+
+        if (pacienteEncontrado.isPresent()) {
+            pacienteRepository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException("{\"mensaje\": \"Paciente no encontrado\"}");
+        }
     }
 
     @Override
